+++ custom_components/octopus_energy/gas/previous_accumulative_consumption_kwh.py
@@ -157,6 +157,7 @@ class OctopusEnergyPreviousAccumulativeGasConsumptionKwh(CoordinatorEntity, Octo
 
     if result is not None:
       self._attributes["data_last_retrieved"] = result.last_retrieved
+      self._attributes["latest_available_data_timestamp"] = result.latest_available_timestamp
 
   async def async_added_to_hass(self):
     """Call when entity about to be added to hass."""