+++ custom_components/octopus_energy/sensor.py
@@ -4,6 +4,7 @@ import logging
 from homeassistant.util.dt import (utcnow)
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_platform, issue_registry as ir, entity_registry as er
+import homeassistant.helpers.config_validation as cv
 
 from .electricity.current_consumption import OctopusEnergyCurrentElectricityConsumption
 from .electricity.current_accumulative_consumption import OctopusEnergyCurrentAccumulativeElectricityConsumption
@@ -41,7 +42,14 @@ from .wheel_of_fortune.gas_spins import OctopusEnergyWheelOfFortuneGasSpins
 from .cost_tracker.cost_tracker import OctopusEnergyCostTrackerSensor
 from .cost_tracker.cost_tracker_off_peak import OctopusEnergyCostTrackerOffPeakSensor
 from .cost_tracker.cost_tracker_peak import OctopusEnergyCostTrackerPeakSensor
+from .cost_tracker.cost_tracker_week import OctopusEnergyCostTrackerWeekSensor
+from .cost_tracker.cost_tracker_week_off_peak import OctopusEnergyCostTrackerWeekOffPeakSensor
+from .cost_tracker.cost_tracker_week_peak import OctopusEnergyCostTrackerWeekPeakSensor
+from .cost_tracker.cost_tracker_month import OctopusEnergyCostTrackerMonthSensor
+from .cost_tracker.cost_tracker_month_off_peak import OctopusEnergyCostTrackerMonthOffPeakSensor
+from .cost_tracker.cost_tracker_month_peak import OctopusEnergyCostTrackerMonthPeakSensor
 from .greenness_forecast.current_index import OctopusEnergyGreennessForecastCurrentIndex
+from .greenness_forecast.next_index import OctopusEnergyGreennessForecastNextIndex
 
 from .coordinators.current_consumption import async_create_current_consumption_coordinator
 from .coordinators.gas_rates import async_setup_gas_rates_coordinator
@@ -75,7 +83,6 @@ from .const import (
   CONFIG_MAIN_GAS_PRICE_CAP,
 
   DATA_ELECTRICITY_RATES_COORDINATOR_KEY,
-  DATA_SAVING_SESSIONS_COORDINATOR,
   DATA_CLIENT,
   DATA_ACCOUNT
 )
@@ -119,7 +126,7 @@ async def async_setup_entry(hass, entry, async_add_entities):
       # supports_response=SupportsResponse.OPTIONAL
     )
   elif config[CONFIG_KIND] == CONFIG_KIND_COST_TRACKER:
-    await async_setup_cost_sensors(hass, config, async_add_entities)
+    await async_setup_cost_sensors(hass, entry, config, async_add_entities)
 
     platform = entity_platform.async_get_current_platform()
     platform.async_register_entity_service(
@@ -133,6 +140,46 @@ async def async_setup_entry(hass, entry, async_add_entities):
         ),
       ),
       "async_update_cost_tracker_config"
+    )
+
+    platform.async_register_entity_service(
+      "reset_cost_tracker",
+      vol.All(
+        vol.Schema(
+          {},
+          extra=vol.ALLOW_EXTRA,
+        ),
+      ),
+      "async_reset_cost_tracker"
+    )
+
+    platform.async_register_entity_service(
+      "adjust_accumulative_cost_tracker",
+      vol.All(
+        vol.Schema(
+          {
+            vol.Required("date"): cv.date,
+            vol.Required("consumption"): cv.positive_float,
+            vol.Required("cost"): cv.positive_float,
+          },
+          extra=vol.ALLOW_EXTRA,
+        ),
+      ),
+      "async_adjust_accumulative_cost_tracker"
+    )
+
+    platform.async_register_entity_service(
+      "adjust_cost_tracker",
+      vol.All(
+        vol.Schema(
+          {
+            vol.Required("datetime"): cv.datetime,
+            vol.Required("consumption"): cv.positive_float,
+          },
+          extra=vol.ALLOW_EXTRA,
+        ),
+      ),
+      "async_adjust_cost_tracker"
     )
 
 async def async_setup_default_sensors(hass: HomeAssistant, config, async_add_entities):
@@ -149,10 +196,10 @@ async def async_setup_default_sensors(hass: HomeAssistant, config, async_add_ent
   entities = [
     OctopusEnergyWheelOfFortuneElectricitySpins(hass, wheel_of_fortune_coordinator, client, account_id),
     OctopusEnergyWheelOfFortuneGasSpins(hass, wheel_of_fortune_coordinator, client, account_id),
-    OctopusEnergyGreennessForecastCurrentIndex(hass, greenness_forecast_coordinator, account_id)
+    OctopusEnergyGreennessForecastCurrentIndex(hass, greenness_forecast_coordinator, account_id),
+    OctopusEnergyGreennessForecastNextIndex(hass, greenness_forecast_coordinator, account_id)
   ]
 
-  registry = er.async_get(hass)
   entity_ids_to_migrate = []
 
   if account_info["octoplus_enrolled"] == True:
@@ -282,7 +329,7 @@ async def async_setup_default_sensors(hass: HomeAssistant, config, async_add_ent
             None,
             previous_gas_consumption_days_offset
           )
-          entities.append(OctopusEnergyPreviousAccumulativeGasConsumptionCubicMeters(hass, client, previous_consumption_coordinator, meter, point, calorific_value))
+          entities.append(OctopusEnergyPreviousAccumulativeGasConsumptionCubicMeters(hass, client, previous_consumption_coordinator, account_id, meter, point, calorific_value))
           entities.append(OctopusEnergyPreviousAccumulativeGasConsumptionKwh(hass, previous_consumption_coordinator, meter, point, calorific_value))
           entities.append(OctopusEnergyPreviousAccumulativeGasCost(hass, previous_consumption_coordinator, meter, point, calorific_value))
           entities.append(OctopusEnergyPreviousAccumulativeGasCostOverride(hass,  account_id, previous_consumption_coordinator, client, gas_tariff_code, meter, point, calorific_value))
@@ -340,28 +387,43 @@ async def async_setup_default_sensors(hass: HomeAssistant, config, async_add_ent
 
   async_add_entities(entities)
 
-async def async_setup_cost_sensors(hass: HomeAssistant, config, async_add_entities):
+async def async_setup_cost_sensors(hass: HomeAssistant, entry, config, async_add_entities):
   account_id = config[CONFIG_ACCOUNT_ID]
   account_result = hass.data[DOMAIN][account_id][DATA_ACCOUNT]
   account_info = account_result.account if account_result is not None else None
 
   mpan = config[CONFIG_COST_MPAN]
 
+  registry = er.async_get(hass)
+
   now = utcnow()
-  is_export = False
   for point in account_info["electricity_meter_points"]:
     tariff_code = get_active_tariff_code(now, point["agreements"])
     if tariff_code is not None:
       # For backwards compatibility, pick the first applicable meter
       if point["mpan"] == mpan or mpan is None:
         for meter in point["meters"]:
-          is_export = meter["is_export"]
           serial_number = meter["serial_number"]
           coordinator = hass.data[DOMAIN][account_id][DATA_ELECTRICITY_RATES_COORDINATOR_KEY.format(mpan, serial_number)]
+
+          sensor = OctopusEnergyCostTrackerSensor(hass, coordinator, config)
+          off_peak_sensor = OctopusEnergyCostTrackerOffPeakSensor(hass, coordinator, config)
+          peak_sensor = OctopusEnergyCostTrackerPeakSensor(hass, coordinator, config)
+
+          sensor_entity_id = registry.async_get_entity_id("sensor", DOMAIN, sensor.unique_id)
+          off_peak_sensor_entity_id = registry.async_get_entity_id("sensor", DOMAIN, off_peak_sensor.unique_id)
+          peak_sensor_entity_id = registry.async_get_entity_id("sensor", DOMAIN, peak_sensor.unique_id)
+
           entities = [
-            OctopusEnergyCostTrackerSensor(hass, coordinator, config, is_export),
-            OctopusEnergyCostTrackerOffPeakSensor(hass, coordinator, config, is_export),
-            OctopusEnergyCostTrackerPeakSensor(hass, coordinator, config, is_export)
+            sensor,
+            off_peak_sensor,
+            peak_sensor,
+            OctopusEnergyCostTrackerWeekSensor(hass, entry, config, sensor_entity_id if sensor_entity_id is not None else sensor.entity_id),
+            OctopusEnergyCostTrackerWeekOffPeakSensor(hass, entry, config, off_peak_sensor_entity_id if off_peak_sensor_entity_id is not None else off_peak_sensor.entity_id),
+            OctopusEnergyCostTrackerWeekPeakSensor(hass, entry, config, peak_sensor_entity_id if peak_sensor_entity_id is not None else peak_sensor.entity_id),
+            OctopusEnergyCostTrackerMonthSensor(hass, entry, config, sensor_entity_id if sensor_entity_id is not None else sensor.entity_id),
+            OctopusEnergyCostTrackerMonthOffPeakSensor(hass, entry, config, off_peak_sensor_entity_id if off_peak_sensor_entity_id is not None else off_peak_sensor.entity_id),
+            OctopusEnergyCostTrackerMonthPeakSensor(hass, entry, config, peak_sensor_entity_id if peak_sensor_entity_id is not None else peak_sensor.entity_id),
           ]
           async_add_entities(entities)
-          return
+          break